requirements.txt
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
h11==0.14.0
idna==3.7
imagedominantcolor==1.0.1
kiwisolver==1.4.5
matlab==0.1
matplotlib==3.9.2
numpy==2.0.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.14.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
